tests: add a script to create a test vm
To test the kernel installation and booting, we need a remote peer
running a ssh server.

We add a script that creates a VM which we can install the kernel into.
To connect to the VM ssh, we need to run the following command after
we start the VM:

    virsh qemu-monitor-command --hmp pyktest 'hostfwd_add ::2222-:22'

Then we can connect to VM through:

    ssh -p 2222 root@localhost

The password for root is 123456.